requirements.txt
fastapi
uvicorn
pydantic
requests
pytest
httpx
python-dotenv
sqlmodel
sqlalchemy
pillow
